+++ motion/__init__.py
@@ -29,13 +29,13 @@
 
 """
 
-motion is a Django application for community microblogging through the TypePad
+Motion is a Django application for community microblogging through the TypePad
 API.
 
 """
 
-__version__ = '1.1.1'
-__date__ = '16 December 2009'
+__version__ = '1.1.2'
+__date__ = '30 March 2010'
 __author__ = 'Six Apart Ltd.'
 __credits__ = """Brad Choate
 Leah Culver